Fix a bug in AnimatedSwitch (#3)
Remove the unused property onStateChange on the TransitionRoute, child of
AnimatedSwitch. Also remove the need to clone the child being returned by
render.

Fix the test for AnimatedSwitch which now passes correctly with the full
cycle, including WILL_ENTER/DID_ENTER.

Remove the use of this.status in AnimatedSwitch, it now uses only the
state.