Parallel transition (#4)
* Add entering and exiting transition at the same time

Add a parallel property which should be renamed for more clarity. In
order to allow the exiting and entering transitions to run in sync.

Rewrite most of AnimatedSwitch and get rid of TransitionRoute and its
associated TransitionTransmitter.

Rename AnimatedSwitch in TransitionSwitch.

* Fix synchronisation Issues

Fix some issues in the synchronization. Like if the route changes while
transitioning.

Improve documentation.

Fix tests