Fix child properties issue
In the current release, the props given to the components under the
TransitionSwitch are not dynamic anymore. This is due to the fact that
we store a clone from this.props.children in the internal state of
TransitionSwitch.

The current solution is to only store the key in the state, and in the
render method we retrive in this.props.children the ones that match the
keys from the state, and we render these