Fix prop issue (#8)
* Fix child properties issue

In the current release, the props given to the components under the
TransitionSwitch are not dynamic anymore. This is due to the fact that
we store a clone from this.props.children in the internal state of
TransitionSwitch.

The current solution is to only store the key in the state, and in the
render method we retrive in this.props.children the ones that match the
keys from the state, and we render these

* Add support for render and component props of <Route />

Add support for <Route/>'s component prop. It is the most common use
case and it is now supported.

Add support for <Route/>'s render prop. It is now going to fully support
the react router v4 Route API.

TransitionSwitch no longer returns the underlying Route, but only what
the Route element would render, meaning the child given to the Route
via component, render, or children.