Add ShouldComponentUpdate implementation
Add ShouldComponentUpdate. It fixes the problem where changing to a url that would
be resolved by the same Route as the previous one. eg: a Route with a path such as
/path/:param where you would go from /path/1 to /path/2. Previously the underlying
component would unmount and remount. Now the transition only occurs when a new
Route is being resolved, much like how the original TransitionGroup works.